pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "market-signal-lab"
version = "0.1.0"
description = "Local-first trading research, backtesting, and monitoring for US/Canada stocks and crypto"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "duckdb>=1.0",
    "pandas>=2.1",
    "numpy>=1.24",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "jinja2>=3.1",
    "python-multipart>=0.0.9",
    "plotly>=5.20",
    "pyarrow>=15.0",
    "click>=8.1",
    "rich>=13.7",
    "tenacity>=8.2",
]

[project.optional-dependencies]
ml = [
    "scikit-learn>=1.4",
    "joblib>=1.3",
]
ai = [
    "ollama>=0.2",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "httpx>=0.27",
]

[project.scripts]
msl = "app.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"
